Converted the whole thing into a single-file project
