--- tags/alltags.h
@@ -1,11 +1,0 @@
-#pragma once
-#include "compoundtag.h"
-#include "listtag.h"
-#include "stringtag.h"
-#include "bytetag.h"
-#include "shorttag.h"
-#include "inttag.h"
-#include "longtag.h"
-#include "floattag.h"
-#include "doubletag.h"
-#include "bytearraytag.h"
